[Refactor] Migrate StreamLoadPipe to ComputeEnv
