+++ be/src/compute_env/CMakeLists.txt
@@ -89,6 +89,10 @@ set(COMPUTE_ENV_LOAD_PATH_FILES
     load_path/rejected_record_writer.cpp
 )
 
+set(COMPUTE_ENV_LOAD_FILES
+    load/stream_load_pipe.cpp
+)
+
 set(COMPUTE_ENV_STAROS_FILES
     staros/staros_status.cpp
     staros/staros_starcache.cpp
@@ -122,6 +126,7 @@ ADD_BE_LIB(ComputeEnv
     ${COMPUTE_ENV_GLOBAL_DICT_FILES}
     ${COMPUTE_ENV_DICTIONARY_CACHE_FILES}
     ${COMPUTE_ENV_LOAD_PATH_FILES}
+    ${COMPUTE_ENV_LOAD_FILES}
     ${COMPUTE_ENV_STAROS_FILES}
 )
