+++ be/src/compute_env/load/stream_load_pipe.cpp
@@ -32,7 +32,7 @@
 // specific language governing permissions and limitations
 // under the License.
 
-#include "runtime/stream_load/stream_load_pipe.h"
+#include "compute_env/load/stream_load_pipe.h"
 
 #include "base/compression/compression_utils.h"
 #include "base/utility/alignment.h"
