+++ be/src/exec/file_scanner/avro_scanner.h
@@ -25,7 +25,7 @@
 #include "exec/file_scanner/json_scanner.h"
 #include "exprs/json_functions.h"
 #include "fs/fs.h"
-#include "runtime/stream_load/load_stream_mgr.h"
+#include "compute_env/load/stream_load_pipe.h"
 #ifdef __cplusplus
 extern "C" {
 #endif
