+++ be/src/http/action/transaction_stream_load.cpp
@@ -57,7 +57,7 @@
 #include "runtime/stream_load/load_stream_mgr.h"
 #include "runtime/stream_load/stream_load_context.h"
 #include "runtime/stream_load/stream_load_executor.h"
-#include "runtime/stream_load/stream_load_pipe.h"
+#include "compute_env/load/stream_load_pipe.h"
 #include "runtime/stream_load/transaction_mgr.h"
 
 namespace starrocks {
