+++ be/src/runtime/CMakeLists.txt
@@ -19,6 +19,7 @@ set(EXECUTABLE_OUTPUT_PATH "${BUILD_DIR}/src/runtime")
 
 set(RUNTIME_CORE_FILES
     byte_buffer.cpp
+    message_body_sink.cpp
     current_thread.cpp
     chunk_helper.cpp
     descriptors.cpp
@@ -95,12 +96,10 @@ set(RUNTIME_FILES
     load_channel.cpp
     local_tablets_channel.cpp
     snapshot_loader.cpp
-    message_body_sink.cpp
     stream_load/stream_load_context_handle.cpp
     stream_load/transaction_mgr.cpp
     stream_load/stream_load_context.cpp
     stream_load/stream_load_executor.cpp
-    stream_load/stream_load_pipe.cpp
     stream_load/time_bounded_stream_load_pipe.cpp
     batch_write/isomorphic_batch_write.cpp
     batch_write/batch_write_mgr.cpp
