+++ be/src/runtime/routine_load/kafka_consumer_pipe.h
@@ -42,7 +42,7 @@
 #include "librdkafka/rdkafka.h"
 #include "pulsar/Client.h"
 #include "runtime/message_body_sink.h"
-#include "runtime/stream_load/stream_load_pipe.h"
+#include "compute_env/load/stream_load_pipe.h"
 #include "simdjson.h"
 
 namespace starrocks {
