+++ be/src/runtime/stream_load/transaction_mgr.cpp
@@ -55,7 +55,7 @@
 #include "runtime/stream_load/stream_load_context.h"
 #include "runtime/stream_load/stream_load_executor.h"
 #include "runtime/stream_load/stream_load_metrics.h"
-#include "runtime/stream_load/stream_load_pipe.h"
+#include "compute_env/load/stream_load_pipe.h"
 #include "runtime/stream_load/transaction_mgr.h"
 
 namespace starrocks {
