+++ be/src/runtime/stream_load/transaction_mgr.h
@@ -23,7 +23,7 @@
 #include "fmt/format.h"
 #include "runtime/exec_env.h"
 #include "runtime/stream_load/load_stream_mgr.h"
-#include "runtime/stream_load/stream_load_pipe.h"
+#include "compute_env/load/stream_load_pipe.h"
 #include "stream_load_context.h"
 
 namespace starrocks {
