+++ be/test/exec/connector_scan_node_test.cpp
@@ -29,7 +29,7 @@
 #include "runtime/exec_env.h"
 #include "runtime/runtime_state.h"
 #include "runtime/stream_load/load_stream_mgr.h"
-#include "runtime/stream_load/stream_load_pipe.h"
+#include "compute_env/load/stream_load_pipe.h"
 #include "storage/query/olap_morsel_queue.h"
 
 namespace starrocks {
