+++ be/test/exec/file_scanner/json_scanner_test.cpp
@@ -36,7 +36,7 @@
 #include "runtime/descriptors.h"
 #include "runtime/exec_env.h"
 #include "runtime/runtime_state.h"
-#include "runtime/stream_load/stream_load_pipe.h"
+#include "compute_env/load/stream_load_pipe.h"
 
 namespace starrocks {
